Minor bug fixes, unused code removals. Unit Tests for business.pet.

+++ hpf-api/src/main/java/fontys/sem3/hpfapi/business/pet/impl/DeletePetUseCaseImpl.java
@@ -1,6 +1,8 @@
 package fontys.sem3.hpfapi.business.pet.impl;
 
+import fontys.sem3.hpfapi.business.exception.UnauthorizedDataAccessException;
 import fontys.sem3.hpfapi.business.pet.DeletePetUseCase;
+import fontys.sem3.hpfapi.dto.login.AccessTokenDTO;
 import fontys.sem3.hpfapi.repository.PetRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -11,10 +13,15 @@ import javax.transaction.Transactional;
 @RequiredArgsConstructor
 public class DeletePetUseCaseImpl implements DeletePetUseCase {
     private final PetRepository petRepository;
+    private final AccessTokenDTO requestAccessToken;
 
     @Transactional
     @Override
     public void deletePet(long petId) {
-        this.petRepository.deleteById(petId);
+        if (requestAccessToken.hasRole("ADMIN")) {
+            this.petRepository.deleteById(petId);
+        } else {
+            throw new UnauthorizedDataAccessException("ACCESS_DENIED");
+        }
     }
 }
